# api/user/src/adapters/controllers/openapi/generated/api/openapi.yaml
openapi: 3.0.0
info:
  title: User
  version: '1.0'
servers:
  - description: local
    url: http://localhost:3001
  - description: dev
    url: https://api-user-dot-horai-dev-scheme-verge-v2.an.r.appspot.com
paths:
...
  /amos/{amoId}:
    get:
      operationId: getAmoById
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: amoId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmoBody'
          description: OK
        '404':
          description: Not Found
      security:
        - FirebaseAuthentication: []
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
    put:
      operationId: amosAmoIdPUT
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: amoId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmoBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
  /amos:
    get:
      operationId: getAmos
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Amo'
                type: array
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get AMOs
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
  /v2/amos/deactivate/{amoId}:
    put:
      operationId: updateAmoStatus
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: amoId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmoStatusBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /v2/amos/{amoId}:
    delete:
      operationId: deleteAmoByamoId
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: amoId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: deletes amo by amoId
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
    get:
      operationId: getAmoByamoId
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: amoId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AmoBody'
          description: OK
        '404':
          description: Not Found
      security:
        - FirebaseAuthentication: []
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
    put:
      operationId: updateAmobyamoId
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: amoId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmoBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /v2/amos:
    get:
      operationId: getAllAmos
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Amo'
                type: array
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get All AMOs
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
    post:
      operationId: createAmo
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmoBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /smbs:
    get:
      operationId: getSmbs
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: true
          in: query
          name: amoId
          required: false
          schema:
            type: string
          style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Smb'
                type: array
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get SMBs
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
  /smbs/{smbId}:
    get:
      operationId: getSmbById
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Smb'
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get SMB by id
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
    put:
      operationId: updateSmb
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmbBody'
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Update SMB
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
  /v2/smb/deactivate/{smbId}:
    put:
      operationId: deactivateSmb
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmbStatusBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /v2/smbs:
    get:
      operationId: getAllSmbs
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Smb'
                type: array
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get All SMBs
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
    post:
      operationId: createSmb
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmbBody'
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Create SMB
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /v2/smbs/{smbId}:
    get:
      operationId: getSmbBySmbId
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Smb'
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get v2 SMB by id
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
    put:
      operationId: updateSmbBySmbId
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmbUpdateBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Update Smb
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /v2/smbs/{smbId}/reports:
    get:
      operationId: getReportsBySmbId
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Report'
                type: array
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Get v2 SMB reports by id
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /v2/smbs/{smbId}/reports/{reportId}:
    put:
      operationId: updateReport
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: Id of specific Smb
          explode: false
          in: path
          name: smbId
          required: true
          schema:
            type: string
          style: simple
        - description: Id of specific Report
          explode: false
          in: path
          name: reportId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportUpdateBody'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Update Report
      tags:
        - user
      x-eov-operation-handler: controllers/UserController
  /coWorkingSpace/theSeatHalki:
    get:
      operationId: getTheSeatHalkiSmb
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Smb'
          description: OK
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      summary: Get The Seat Halki detail
      tags:
        - coWorkingSpaceSmb
      x-eov-operation-handler: controllers/CoWorkingSpaceSmbController
  /coWorkingSpace/techHubYokohama:
    get:
      operationId: getTechHubYokohama
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Smb'
          description: OK
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      summary: Get The Tech Hub Yokohama Details
      tags:
        - coWorkingSpaceSmb
      x-eov-operation-handler: controllers/CoWorkingSpaceSmbController
  /users:
    get:
      operationId: listUsers
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - description: start Index, defaut is 1
          explode: true
          in: query
          name: startIndex
          required: false
          schema:
            type: integer
          style: form
        - description: end Index, default is 20
          explode: true
          in: query
          name: endIndex
          required: false
          schema:
            type: integer
          style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
          description: OK
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: get users
      tags:
        - app
      x-eov-operation-handler: controllers/AppController
...
  /customer:
    get:
      operationId: getAllCustomers
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
        - explode: true
          in: query
          name: page
          required: false
          schema:
            type: number
          style: form
        - explode: true
          in: query
          name: limit
          required: false
          schema:
            type: number
          style: form
        - allowReserved: true
          explode: true
          in: query
          name: searchText
          required: false
          schema:
            type: string
          style: form
        - explode: true
          in: query
          name: serviceType
          required: false
          schema:
            enum:
              - Application
              - Reservation
              - CoWorkingSpace
            type: string
          style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'
          description: OK
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: GET all customers
      tags:
        - Customers
      x-eov-operation-handler: controllers/CustomersController
    post:
      operationId: createCustomer
      parameters:
        - description: 使用不可.　この変数は設定してはいけない.　bearer認証用token.　ServerCode生成用.
          explode: false
          in: header
          name: authorization
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerBody'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
          description: OK
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      security:
        - FirebaseAuthentication: []
      summary: Create customer
      tags:
        - Customers
      x-eov-operation-handler: controllers/CustomersController
...
  /customer/generateOtp:
    post:
      operationId: generateOtp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateOtpBody'
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  hash:
                    type: string
                  message:
                    type: string
          description: OK
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
      summary: Generate OTP to be sent on email address provided
      tags:
        - Auth
      x-eov-operation-handler: controllers/AuthController
...
components:
...
  schemas:
...
    AmoBody:
      example:
        name: name
        email: email
      properties:
        name:
          type: string
        email:
          type: string
      required:
        - email
        - name
      type: object
    AmoStatusBody:
      example:
        isActive: true
      properties:
        isActive:
          type: boolean
      required:
        - isActive
      type: object
    Amo:
      allOf:
        - $ref: '#/components/schemas/AmoBody'
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        amoId: amoId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        amoId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - amoId
        - createdAt
        - name
        - updatedAt
      type: object
...
    Smb:
      allOf:
        - $ref: '#/components/schemas/SmbBody'
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        smbId: smbId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        smbId:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - amoId
        - createdAt
        - name
        - smbId
        - updatedAt
      title: SMB
...
  securitySchemes:
    FirebaseAuthentication:
      scheme: bearer
      type: http